Updated comment controllers

+++ controllers/game.js
@@ -9,6 +9,9 @@ const {
   deleteImage,
 } = require("../helpers/controller");
 
+const defaultImage =
+  "https://upload.wikimedia.org/wikipedia/commons/thumb/a/ac/No_image_available.svg/480px-No_image_available.svg.png";
+
 const findAll = (limit = 0, offset = 0) =>
   execController(
     skipNext,
@@ -20,9 +23,8 @@ const findOne = (filter, data) =>
 
 const create = (game, file) =>
   execController(async () => {
-    const image = file
-      ? await uploadImage(file, "1GWaB-McnGh3L1L3ICQkC0ek7o7GMEHg0")
-      : "https://upload.wikimedia.org/wikipedia/commons/thumb/a/ac/No_image_available.svg/480px-No_image_available.svg.png";
+    const gamesFolder = "1GWaB-McnGh3L1L3ICQkC0ek7o7GMEHg0";
+    const image = file ? await uploadImage(file, gamesFolder) : defaultImage;
     await Game.create({ ...game, image });
     await User.findOneAndUpdate(
       { _id: game.user },
@@ -57,11 +59,7 @@ const deleteOne = (filter) =>
     );
     await Comment.deleteMany({ postedOn: "game", postedOnId: _id });
     await Transaction.deleteMany({ game: _id });
-    if (
-      image !==
-      "https://upload.wikimedia.org/wikipedia/commons/thumb/a/ac/No_image_available.svg/480px-No_image_available.svg.png"
-    )
-      await deleteImage(image);
+    if (image !== defaultImage) await deleteImage(image);
   }, Game.deleteOne(filter));
 
 module.exports = {
